src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
alsa-lib=1.2.3=h516909a_0
appdirs=1.4.4=pyh9f0ad1d_0
astroid=2.6.0=py38h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
audioread=2.1.9=py38h578d9bd_0
black=19.10b0=py_0
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7f98852_4
ca-certificates=2021.5.30=ha878542_0
certifi=2021.5.30=py38h578d9bd_0
cffi=1.14.5=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_1
click=8.0.1=pyhd3eb1b0_0
cryptography=3.4.7=py38ha5dfef3_0
cycler=0.10.0=py_2
dbus=1.13.6=h48d8840_2
decorator=5.0.9=pyhd8ed1ab_0
expat=2.4.1=h9c3ff4c_0
ffmpeg=4.3.1=hca11adc_2
flake8=3.9.2=pyhd3eb1b0_0
fluidsynth=v1.1.11=h63f31ab_1001
fontconfig=2.13.1=hba837de_1005
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h0b5b191_1005
glib=2.68.3=h9c3ff4c_0
glib-tools=2.68.3=h9c3ff4c_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gst-plugins-base=1.18.4=hf529b03_2
gstreamer=1.18.4=h76c114f_2
icu=68.1=h58526e2_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=4.6.0=py38h578d9bd_0
importlib_metadata=4.6.0=hd8ed1ab_0
isort=5.9.1=pyhd3eb1b0_0
jbig=2.1=h7f98852_2003
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
jsonpickle=2.0.0=pyhd8ed1ab_0
kiwisolver=1.3.1=py38h1fd1430_1
krb5=1.19.1=hcc1bbae_0
lame=3.100=h7f98852_1001
lazy-object-proxy=1.6.0=py38h27cfd23_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_0
lerc=2.2.1=h9c3ff4c_0
libblas=3.9.0=9_openblas
libcblas=3.9.0=9_openblas
libclang=11.1.0=default_ha53f305_1
libdeflate=1.7=h7f98852_5
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=hcdb4288_3
libffi=3.3=h58526e2_2
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=9.3.0=h2828fa1_19
libgfortran-ng=9.3.0=hff62375_19
libgfortran5=9.3.0=hff62375_19
libglib=2.68.3=h3e27bee_0
libgomp=9.3.0=h2828fa1_19
libiconv=1.16=h516909a_0
liblapack=3.9.0=9_openblas
libllvm10=10.0.1=he513fc3_3
libllvm11=11.1.0=hf817b99_2
libogg=1.3.4=h7f98852_1
libopenblas=0.3.15=pthreads_h8fe5266_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libpq=13.3=hd57d9b9_0
librosa=0.8.1=pyhd8ed1ab_0
libsndfile=1.0.31=h9c3ff4c_1
libstdcxx-ng=9.3.0=h6de172a_19
libtiff=4.3.0=hf544144_1
libuuid=2.32.1=h7f98852_1000
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.0=h7f98852_2
libxcb=1.13=h7f98852_1003
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
llvmlite=0.36.0=py38h4630a5e_0
lz4-c=1.9.3=h9c3ff4c_0
mad=0.15.1b=he1b5a44_0
matplotlib=3.4.2=py38h578d9bd_0
matplotlib-base=3.4.2=py38hcc49a3a_0
mccabe=0.6.1=py38_1
mido=1.2.10=pypi_0
more-itertools=8.8.0=pyhd8ed1ab_0
music21=6.7.1=pyhd8ed1ab_0
muspy=0.3.0=pypi_0
mypy_extensions=0.4.3=py38_0
mysql-common=8.0.25=ha770c72_2
mysql-libs=8.0.25=hfa10184_2
ncurses=6.2=h58526e2_4
nettle=3.6=he412f7d_0
nspr=4.30=h9c3ff4c_0
nss=3.67=hb5efdd6_0
numba=0.53.1=py38h8b71fd7_1
numpy=1.21.0=py38h9894fe3_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1k=h7f98852_0
packaging=20.9=pyh44b312d_0
pathspec=0.7.0=py_0
pcre=8.45=h9c3ff4c_0
pillow=8.2.0=py38ha0e1e83_1
pip=21.1.3=pyhd8ed1ab_0
pooch=1.4.0=pyhd8ed1ab_0
pretty-midi=0.2.9=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.20=pyh9f0ad1d_2
pyflakes=2.3.1=pyhd3eb1b0_0
pylint=2.8.3=py38h06a4308_1
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pypianoroll=1.0.4=pypi_0
pyqt=5.12.3=py38h578d9bd_7
pyqt-impl=5.12.3=py38h7400c14_7
pyqt5-sip=4.19.18=py38h709712a_7
pyqtchart=5.12=py38h7400c14_7
pyqtwebengine=5.12.1=py38h7400c14_7
pysocks=1.7.1=py38h578d9bd_3
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.8.10=h49503c6_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=2_cp38
pyyaml=5.4.1=pypi_0
qt=5.12.9=hda022c4_4
readline=8.1=h46c0cb4_0
regex=2021.4.4=py38h27cfd23_0
requests=2.25.1=pyhd3deb0d_0
resampy=0.2.2=py_0
scikit-learn=0.24.2=py38hdc147b9_0
scipy=1.7.0=py38h7b17777_0
setuptools=49.6.0=py38h578d9bd_3
six=1.16.0=pyh6c4a22f_0
sox=14.4.2=h27f72bc_1013
sqlite=3.36.0=h9cd32fc_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h21135ba_1
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py38h497a2fe_1
tqdm=4.61.2=pyhd8ed1ab_0
typed-ast=1.4.3=py38h7f8727e_1
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
webcolors=1.11.1=pyhd8ed1ab_0
wheel=0.36.2=pyhd3deb0d_0
wrapt=1.12.1=py38h7b6447c_1
x264=1!161.3030=h7f98852_1
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h516909a_1010
zstd=1.5.0=ha95c52a_0
